Rely solely on PORTAL_CONFIG for application startup

+++ ROOT/opt/supervisor-scripts/caddy.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-[[ ${AUTOSCALER,,} = 'true' ]] && echo "Refusing to start ${PROC_NAME} (AUTOSCALER=true)" | tee -a "/var/log/portal/${PROC_NAME}.log" && exit
-[[ $OPEN_BUTTON_PORT != "1111" ]] && echo "Refusing to start ${PROC_NAME} (OPEN_BUTTON_PORT!=1111)" | tee -a "/var/log/portal/${PROC_NAME}.log" && exit
-
 # Run the caddy configurator
 cd /opt/portal-aio/caddy_manager
 /opt/portal-aio/venv/bin/python caddy_config_manager.py | tee -a "/var/log/portal/${PROC_NAME}.log"
 
-# Frontend log viewer will force a page reload if this string is detected
-echo "Starting Caddy..." | tee -a "/var/log/portal/${PROC_NAME}.log"
+# Ensure the portal config file exists if running without PORTAL_CONFIG
+touch /etc/portal.yaml
 
-/opt/portal-aio/caddy_manager/caddy run --config /etc/Caddyfile | tee -a "/var/log/portal/${PROC_NAME}.log"
+if [[ -f /etc/Caddyfile ]]; then
+    # Frontend log viewer will force a page reload if this string is detected
+    echo "Starting Caddy..." | tee -a "/var/log/portal/${PROC_NAME}.log"
+    /opt/portal-aio/caddy_manager/caddy run --config /etc/Caddyfile | tee -a "/var/log/portal/${PROC_NAME}.log"
+else
+    echo "Not Starting Caddy - No config file was generated" | tee -a "/var/log/portal/${PROC_NAME}.log"
+fi
